remove more deprecated code
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1884263 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFRichTextString.java
@@ -168,7 +168,7 @@ public final class HSSFRichTextString implements Comparable<HSSFRichTextString>,
      * @param font          The index of the font to use.
      */
     public void applyFont(int startIndex, int endIndex, Font font) {
-        applyFont(startIndex, endIndex, font.getIndex());
+        applyFont(startIndex, endIndex, (short)font.getIndex());
     }
 
     /**
